backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
chromadb>=0.4.22
sentence-transformers>=2.3.1
groq>=0.4.2
pymupdf>=1.23.0
python-docx>=1.1.0
python-multipart>=0.0.9
pydantic>=2.5.0
python-dotenv>=1.0.0
torch>=2.1.0
